refactor(session): Rename and apply vaccel type changes
Update implementation to apply vaccel naming and type changes for
session object:
- Update operations to use renamed `vaccel_session_*()` functions
- Change session id type to vaccel_id_t/VaccelId where needed
- Rename session client functions with the `vaccel_rpc_client` prefix

PR: https://github.com/nubificus/vaccel-rust/pull/5
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>

+++ vaccel-rpc-client/src/ops/image.rs
@@ -14,7 +14,7 @@ use vaccel_rpc_proto::image::ImageClassificationRequest;
 use vaccel_rpc_proto::sync::agent_ttrpc::RpcAgentClient;
 
 impl VaccelRpcClient {
-    pub fn image_classify(&self, sess_id: u32, img: Vec<u8>) -> Result<Vec<u8>> {
+    pub fn image_classify(&self, sess_id: i64, img: Vec<u8>) -> Result<Vec<u8>> {
         let ctx = ttrpc::context::Context::default();
         let req = ImageClassificationRequest {
             session_id: sess_id,
@@ -37,7 +37,7 @@ impl VaccelRpcClient {
 #[no_mangle]
 pub unsafe extern "C" fn image_classify(
     client_ptr: *const VaccelRpcClient,
-    sess_id: u32,
+    sess_id: i64,
     img: *const c_uchar,
     img_len: usize,
     tags: *mut c_uchar,
